PCHR-1798: Extract setContractAsLeaveApproverOf to a helper trait
This is going to be useful in order tests, so there's no reason to
keep it only inside the LeaveManager tests
